Respond to data changes
Make it respond to changes in data, with an animated transition
if the position or size has changed; and also to changes in colour
and opacity in the same way.

Internally this is done by creating a new Circle component representing
a circle, which has observers for the state properties we want to animate.